+++ cmd/kubeadm/BUILD
@@ -36,7 +36,6 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//cmd/kubeadm/app:all-srcs",
-        "//cmd/kubeadm/test:all-srcs",
     ],
     tags = ["automanaged"],
 )
