+++ cmd/kubeadm/app/cmd/alpha/BUILD
@@ -62,10 +62,10 @@ go_test(
         "//cmd/kubeadm/app/constants:go_default_library",
         "//cmd/kubeadm/app/phases/certs:go_default_library",
         "//cmd/kubeadm/app/phases/kubeconfig:go_default_library",
+        "//cmd/kubeadm/app/testing:go_default_library",
+        "//cmd/kubeadm/app/testing/cmd:go_default_library",
+        "//cmd/kubeadm/app/testing/kubeconfig:go_default_library",
         "//cmd/kubeadm/app/util/pkiutil:go_default_library",
-        "//cmd/kubeadm/test:go_default_library",
-        "//cmd/kubeadm/test/cmd:go_default_library",
-        "//cmd/kubeadm/test/kubeconfig:go_default_library",
         "//staging/src/k8s.io/client-go/tools/clientcmd:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
     ],
