+++ cmd/kubeadm/app/phases/etcd/BUILD
@@ -13,9 +13,9 @@ go_test(
     deps = [
         "//cmd/kubeadm/app/apis/kubeadm:go_default_library",
         "//cmd/kubeadm/app/constants:go_default_library",
+        "//cmd/kubeadm/app/testing:go_default_library",
         "//cmd/kubeadm/app/util/etcd:go_default_library",
         "//cmd/kubeadm/app/util/staticpod:go_default_library",
-        "//cmd/kubeadm/test:go_default_library",
         "//vendor/github.com/lithammer/dedent:go_default_library",
     ],
 )
