+++ cmd/kubeadm/app/testing/BUILD
@@ -1,14 +1,10 @@
-package(default_visibility = ["//visibility:public"])
-
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-)
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
     srcs = ["util.go"],
-    importpath = "k8s.io/kubernetes/cmd/kubeadm/test",
+    importpath = "k8s.io/kubernetes/cmd/kubeadm/app/testing",
+    visibility = ["//visibility:public"],
     deps = [
         "//cmd/kubeadm/app/apis/kubeadm:go_default_library",
         "//cmd/kubeadm/app/apis/kubeadm/v1beta2:go_default_library",
@@ -31,9 +27,10 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//cmd/kubeadm/test/cmd:all-srcs",
-        "//cmd/kubeadm/test/kubeconfig:all-srcs",
-        "//cmd/kubeadm/test/resources:all-srcs",
+        "//cmd/kubeadm/app/testing/cmd:all-srcs",
+        "//cmd/kubeadm/app/testing/kubeconfig:all-srcs",
+        "//cmd/kubeadm/app/testing/resources:all-srcs",
     ],
     tags = ["automanaged"],
+    visibility = ["//visibility:public"],
 )
