+++ cmd/kubeadm/app/testing/cmd/BUILD
@@ -1,7 +1,11 @@
-sh_library(
-    name = "legacy-script",
-    srcs = glob(["*.sh"]),
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["util.go"],
+    importpath = "k8s.io/kubernetes/cmd/kubeadm/app/testing/cmd",
     visibility = ["//visibility:public"],
+    deps = ["//vendor/github.com/spf13/cobra:go_default_library"],
 )
 
 filegroup(
@@ -13,10 +17,7 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [
-        ":package-srcs",
-        "//test/cmd/kubeadm:all-srcs",
-    ],
+    srcs = [":package-srcs"],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
 )
