+++ cmd/kubeadm/app/testing/resources/cluster_status.go
