+++ test/cmd/kubeadm/BUILD
@@ -9,11 +9,8 @@ load(
 go_library(
     name = "go_default_library",
     srcs = ["util.go"],
-    importpath = "k8s.io/kubernetes/cmd/kubeadm/test/cmd",
-    deps = [
-        "//vendor/github.com/pkg/errors:go_default_library",
-        "//vendor/github.com/spf13/cobra:go_default_library",
-    ],
+    importpath = "k8s.io/kubernetes/test/cmd/kubeadm",
+    deps = ["//vendor/github.com/pkg/errors:go_default_library"],
 )
 
 go_test(
@@ -25,17 +22,15 @@ go_test(
         "token_test.go",
         "version_test.go",
     ],
-    data = ["//cmd/kubeadm"] + glob(["testdata/**"]),
+    data = glob(["testdata/**"]),
     embed = [":go_default_library"],
-    tags = [
-        "integration",
-        "skip",
-    ],
+    rundir = ".",
+    tags = ["integration"],
     deps = [
         "//cmd/kubeadm/app/constants:go_default_library",
         "//cmd/kubeadm/app/phases/certs:go_default_library",
+        "//cmd/kubeadm/app/testing:go_default_library",
         "//cmd/kubeadm/app/util/pkiutil:go_default_library",
-        "//cmd/kubeadm/test:go_default_library",
         "//vendor/github.com/lithammer/dedent:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/sigs.k8s.io/yaml:go_default_library",
