+++ test/cmd/kubeadm/completion_test.go
