+++ test/cmd/kubeadm/testdata/init/current-component-config.yaml
