+++ test/cmd/kubeadm/testdata/init/old-component-config.yaml
