+++ test/cmd/kubeadm/testdata/init/v1alpha1.yaml
