+++ test/cmd/kubeadm/testdata/init/v1alpha2.yaml
