+++ test/cmd/kubeadm/testdata/init/v1alpha3.yaml
