+++ test/cmd/kubeadm/testdata/init/v1beta1.yaml
