+++ test/cmd/kubeadm/testdata/init/v1beta2.yaml
