+++ test/cmd/kubeadm/token_test.go
