+++ test/cmd/kubeadm/version_test.go
